Fixed build warnings on gcc

+++ GameOpenGL/ShaderManager.cpp.inl
@@ -518,7 +518,7 @@ std::set<std::string> ShaderManager<Traits>::ExtractParameterNames(GameOpenGLSha
             nameBuffer);
         CheckOpenGLError();
 
-        if (nameLength < sizeof(ParamPrefix) || strncmp(nameBuffer, ParamPrefix, sizeof(ParamPrefix)))
+        if (static_cast<size_t>(nameLength) < sizeof(ParamPrefix) || strncmp(nameBuffer, ParamPrefix, sizeof(ParamPrefix)))
         {
             throw GameException("Uniform name \"" + std::string(nameBuffer, nameLength) + "\" does not follow the expected name structure: missing \"param\" prefix");
         }
